Correctly run all pending tasks for EmbeddedChannel when the Channel was closed.
Motivation:

When a user called ctx.close() and used the EmbeddedChannel we did not correctly run all pending tasks which means channelInactive was never called.

Modifications:

Ensure we call closeFuture.setClosed() after we scheduled the channelInactive call. This was we can just attach a ChannelFutureListener to the closeFuture when using the EmbeddedChannel which will take care of running all pending tasks.

Result:

Fixes [#6894].